+++ examples/src/main/java/example/ScalaQuery.scala
@@ -24,27 +24,36 @@ package example
 import akka.actor.ActorSystem
 import akka.stream.scaladsl.Sink
 import com.influxdb.client.scala.InfluxDBClientScalaFactory
+import com.influxdb.query.FluxRecord
 
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-object InfluxDB2ScalaExampleRaw {
+object ScalaQuery {
 
-  implicit val system: ActorSystem = ActorSystem("it-tests")
+  implicit val system: ActorSystem = ActorSystem("examples")
 
   def main(args: Array[String]): Unit = {
+
     val influxDBClient = InfluxDBClientScalaFactory
       .create("http://localhost:8086", "my-token".toCharArray, "my-org")
 
     val fluxQuery = ("from(bucket: \"my-bucket\")\n"
-      + " |> range(start: -5m)"
-      + " |> filter(fn: (r) => (r[\"_measurement\"] == \"cpu\" and r[\"_field\"] == \"usage_system\"))"
-      + " |> sample(n: 5, pos: 1)")
+      + " |> range(start: -1d)"
+      + " |> filter(fn: (r) => (r[\"_measurement\"] == \"cpu\" and r[\"_field\"] == \"usage_system\"))")
 
     //Result is returned as a stream
-    val sink = influxDBClient.getQueryScalaApi().queryRaw(fluxQuery)
+    val results = influxDBClient.getQueryScalaApi().query(fluxQuery)
+
+    //Example of additional result stream processing on client side
+    val sink = results
+      //filter on client side using `filter` built-in operator
+      .filter(it => "cpu0" == it.getValueByKey("cpu"))
+      //take first 20 records
+      .take(20)
       //print results
-      .runWith(Sink.foreach[String](it => println(s"Line: $it")))
+      .runWith(Sink.foreach[FluxRecord](it => println(s"Measurement: ${it.getMeasurement}, value: ${it.getValue}")
+      ))
 
     // wait to finish
     Await.result(sink, Duration.Inf)