+++ examples/src/main/java/example/ScalaQueryDSL.scala
@@ -21,39 +21,42 @@
  */
 package example
 
+import java.time.temporal.ChronoUnit
+
 import akka.actor.ActorSystem
 import akka.stream.scaladsl.Sink
 import com.influxdb.client.scala.InfluxDBClientScalaFactory
 import com.influxdb.query.FluxRecord
+import com.influxdb.query.dsl.Flux
+import com.influxdb.query.dsl.functions.restriction.Restrictions
 
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-object InfluxDB2ScalaExample {
+object ScalaQueryDSL {
 
-  implicit val system: ActorSystem = ActorSystem("it-tests")
+  implicit val system: ActorSystem = ActorSystem("examples")
 
   def main(args: Array[String]): Unit = {
 
     val influxDBClient = InfluxDBClientScalaFactory
       .create("http://localhost:8086", "my-token".toCharArray, "my-org")
 
-    val fluxQuery = ("from(bucket: \"my-bucket\")\n"
-      + " |> range(start: -1d)"
-      + " |> filter(fn: (r) => (r[\"_measurement\"] == \"cpu\" and r[\"_field\"] == \"usage_system\"))")
+    val mem = Flux.from("my-bucket")
+      .range(-30L, ChronoUnit.MINUTES)
+      .filter(Restrictions.and(Restrictions.measurement().equal("mem"), Restrictions.field().equal("used_percent")))
 
     //Result is returned as a stream
-    val results = influxDBClient.getQueryScalaApi().query(fluxQuery)
+    val results = influxDBClient.getQueryScalaApi().query(mem.toString())
 
     //Example of additional result stream processing on client side
     val sink = results
       //filter on client side using `filter` built-in operator
-      .filter(it => "cpu0" == it.getValueByKey("cpu"))
+      .filter(it => it.getValue.asInstanceOf[Double] > 55)
       //take first 20 records
       .take(20)
       //print results
-      .runWith(Sink.foreach[FluxRecord](it => println(s"Measurement: ${it.getMeasurement}, value: ${it.getValue}")
-      ))
+      .runWith(Sink.foreach[FluxRecord](it => println(s"Measurement: ${it.getMeasurement}, value: ${it.getValue}")))
 
     // wait to finish
     Await.result(sink, Duration.Inf)